Restructure TicketVC
